Fixing ServiceString converter return value and changing it in retrofitServiceModule so it works
